pyproject.toml
[project]
name = "gibson"
version = "0.1.0"
description = "A crud app for expenses"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "shell (>=1.0.1,<2.0.0)",
    "djoser (>=2.3.3,<3.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "logbook (>=1.9.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest-django (>=4.11.1,<5.0.0)",
    "black (>=25.11.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
